Added a bunch more fail safes

+++ Project_Shopping_App/User.cs
@@ -130,12 +130,17 @@ namespace shoppingApp
         }
         public override void ViewStoreProducts(double threshold)
         {
-            Console.WriteLine("{0,-25} {1,-25} {2,-25} {3,-10} {4,-10}", "ID", "Name", "Category", "Price", "Quantity");
-            foreach (Product p in StoreList.List.Values)
+            if (threshold > 0)
             {
-                if (p.ProductQuantity >= threshold)
-                    p.Print();
+                foreach (Product p in StoreList.List.Values)
+                {
+                    if (p.ProductQuantity >= threshold)
+                        p.Print();
+                }
+                Console.WriteLine("{0,-25} {1,-25} {2,-25} {3,-10} {4,-10}", "ID", "Name", "Category", "Price", "Quantity");
             }
+            else
+                Console.WriteLine("Wrong value entered!");
         }
         private void AddCustomer()
         {
@@ -164,10 +169,20 @@ namespace shoppingApp
                 {
                     Console.Write("Enter choice: ");
                     int uchoice = Convert.ToInt32(Console.ReadLine());
-                    File.Delete(dir[uchoice - 1]);
+                    if (uchoice >= 1 && uchoice <= dir.Length)
+                    {
+                        if(Login.LoadCustomer(dir[uchoice-1]) != null)
+                            File.Delete(dir[uchoice - 1]);
+                        else
+                        {
+                            Console.WriteLine("User is an admin... can't delete admins!");
+                        }
+                    }
+                    else
+                        throw new FormatException();
                     break;
                 }
-                catch
+                catch(FormatException)
                 {
                     idx = 1;
                     Console.WriteLine("Wrong selection!");
@@ -206,25 +221,52 @@ namespace shoppingApp
             Console.Write("Enter ID of product to remove: ");
             string uchoice = Console.ReadLine();
             double quan = 0;
-            if (StoreList.GetProduct(uchoice).ProductQuantity != 0)
+            if (StoreList.GetProduct(uchoice) != null)
             {
-                Console.Write("Enter quantity: ");
-                try
+                if (StoreList.GetProduct(uchoice).ProductQuantity != 0)
                 {
-                    quan = Convert.ToDouble(Console.ReadLine());
+                    Console.Write("Enter quantity: ");
+                    try
+                    {
+                        quan = Convert.ToDouble(Console.ReadLine());
+                        if (quan > StoreList.GetProduct(uchoice).ProductQuantity)
+                        {
+                            Console.WriteLine("Can\'t remove more than available quantity");
+                            throw new FormatException();
+                        }
+                        if (quan < 0)
+                            throw new FormatException();
+                    }
+                    catch (FormatException)
+                    {
+                        Console.WriteLine("Wrong value entered!");
+                        return;
+                    }
                 }
-                catch (FormatException)
+                if (!StoreList.DeleteProduct(uchoice, quan))
                 {
-                    Console.WriteLine("Wrong value entered!");
-                    return;
+                    Console.WriteLine("Wrong choice!");
                 }
+            }else
+            {
+                Console.WriteLine("Product does not exist!");
             }
-            if (!StoreList.DeleteProduct(uchoice, quan))
+
+        }
+        private void Search()
+        {
+            Console.Clear();
+            Console.Write("Enter ProductId to look for: ");
+            string id = Console.ReadLine();
+            if(StoreList.GetProduct(id) !=null)
             {
-                Console.WriteLine("Wrong choice!");
+                Console.WriteLine("{0,-25} {1,-25} {2,-25} {3,-10} {4,-10}", "ID", "Name", "Category", "Price", "Quantity");
+                StoreList.GetProduct(id).Print();
+            }else
+            {
+                Console.WriteLine("Poduct with Id ({0}) does not exist!", id);
             }
         }
-
 
         public override void HandleChoice(string choice)
         {
@@ -246,6 +288,7 @@ namespace shoppingApp
                     DeleteProductFromStore();
                     break;
                 case 5:
+                    Search();
                     break;
                 case 6:
                     ViewAllUsers();
@@ -390,6 +433,8 @@ namespace shoppingApp
                 Product temp = StoreList.GetProduct(tempID).Clone();
                 Console.Write("Enter quantity: ");
                 double uquan = Convert.ToDouble(Console.ReadLine());
+                if (uquan < 0)
+                    throw new FormatException();
                 if (temp.ProductQuantity - uquan >= 0)
                 {
                     temp.ProductQuantity = uquan;
@@ -417,10 +462,10 @@ namespace shoppingApp
         }
         public override void ViewStoreProducts(double threshold = 0)
         {
-            if (threshold == 0)
+            if (threshold >= 0)
                 StoreList.ViewStoreList(false);
         }
-        public void Search()
+        private void Search()
         {
             Dictionary<int, string> categories = new Dictionary<int, string>();
             int idx = 1;
@@ -451,92 +496,113 @@ namespace shoppingApp
 
         private bool Pay()
         {
-            bool success = true;
-            double cost = 0;
-            while (true)
+            if (!Basket.isEmpty())
             {
-                Console.Clear();
-                Basket.ViewBasket();
-                //Calculate cost of items in Basket
-                cost = 0;
-                foreach (Product p in Basket.List.Values)
-                    cost += p.ProductPrice * p.ProductQuantity;
-                if (cost > _balance)
+                double cost = 0;
+                bool success = true;
+                while (true)
                 {
-                    Console.Write("Would you like to remove some products from your basket?(Y/N): ");
-                    string s = Console.ReadLine().ToUpper();
-                    if (s == "Y")
+                    Console.Clear();
+                    Basket.ViewBasket();
+                    Console.WriteLine("Your balance is: {0} ", _balance);
+                    //Calculate cost of items in Basket
+                    cost = 0;
+                    foreach (Product p in Basket.List.Values)
+                        cost += p.ProductPrice * p.ProductQuantity;
+                    if (cost > _balance)
                     {
-                        Console.Write("Enter id of product to remove/remove from: ");
-                        string id = Console.ReadLine();
-                        try
+                        Console.WriteLine("Insufficient funds...");
+                        Console.Write("Would you like to remove some products from your basket?(Y/N): ");
+                        string s = Console.ReadLine().ToUpper();
+                        if (s == "Y")
                         {
-                            Console.Write("Enter quantity to remove: ");
-                            double quan = Convert.ToDouble(Console.ReadLine());
-
-                            if (Basket.DeleteProduct(id, quan))
+                            Console.Write("Enter id of product to remove/remove from: ");
+                            string id = Console.ReadLine();
+                            if(Basket.GetProduct(id) == null)//check if product exists in basket
                             {
-                                StoreList.AddProduct(id, quan); // could be used later if store no longer has a certian item
-                                if (Basket.isEmpty())
+                                Console.WriteLine("Item entered is not in basket!");
+                                Console.Write("Press any key to continue...");
+                                Console.ReadKey();
+                                continue;
+                            }
+                            try
+                            {
+                                Console.Write("Enter quantity to remove: ");
+                                double quan = Convert.ToDouble(Console.ReadLine());
+                                if (quan < 0)
+                                    throw new FormatException();
+                                if (quan != 0)//dont delete if quantity is 0
                                 {
-                                    success = false;
-                                    break;
+                                    if (Basket.DeleteProduct(id, quan))
+                                    {
+                                        StoreList.AddProduct(id, quan);
+                                        if (Basket.isEmpty())
+                                        {
+                                            success = false;
+                                            break;
+                                        }
+                                    }
+                                    else throw new FormatException();
                                 }
                             }
-                            else throw new FormatException(); // id may be not in basket
+                            catch (FormatException)
+                            {
+                                Console.WriteLine("Wrong value entered!");
+                                Console.Write("Press any key to continue...");
+                                Console.ReadKey();
+                            }
                         }
-                        catch (FormatException)
+                        else if (s == "N")
+                        {
+                            success = false;
+                            break;
+                        }
+                        else
                         {
                             Console.WriteLine("Wrong value entered!");
                             Console.Write("Press any key to continue...");
                             Console.ReadKey();
                         }
                     }
-                    else if (s == "N")
-                    {
-                        Console.Write("Press any key to continue...");
-                        success = false;
-                        break;
-                    }
                     else
                     {
-                        Console.WriteLine("Wrong value entered!");
-                        Console.Write("Press any key to continue...");
-                        Console.ReadKey();
+                        success = true;
+                        break;
                     }
                 }
-                else
+                SaveUser();
+                if (success)
                 {
-                    success = true;
-                    break;
+                    while (true)
+                    {
+                        Console.Write("Total cost is {0} and your balance is {1}, would you like to procceed with the transaction?(Y/N): ", cost, _balance);
+                        string uchoice = Console.ReadLine().ToUpper();
+                        if (uchoice == "Y")
+                        {
+                            _balance -= cost;
+                            Basket.List.Clear();
+                            Console.WriteLine("Deducted {0}, new balance is {1}.", cost, _balance);
+                            SaveUser();
+                            break;
+                        }
+                        else if (uchoice == "N")
+                        {
+                            Console.WriteLine("Transaction cancelled.");
+                            break;
+                        }
+                        else
+                        {
+                            Console.WriteLine("Wrong value entered!");
+                        }
+                    }
+                    return true;
                 }
             }
-            SaveUser();
-            if (success)
+            else
             {
-                while (true)
-                {
-                    Console.Write("Total cost is {0} and your balance is {1}, would you like to procceed with the transaction?(Y/N): ", cost, _balance);
-                    string uchoice = Console.ReadLine().ToUpper();
-                    if (uchoice == "Y")
-                    {
-                        _balance -= cost;
-                        Basket.List.Clear();
-                        Console.WriteLine("Deducted {0}, new balance is {1}.", cost, _balance);
-                        SaveUser();
-                        break;
-                    }
-                    else if (uchoice == "N")
-                    {
-                        Console.WriteLine("Transaction cancelled.");
-                        break;
-                    }
-                    else
-                    {
-                        Console.WriteLine("Wrong value entered!");
-                    }
-                }
-                return true;
+                Console.Clear();
+                Basket.ViewBasket();
+                Console.WriteLine("Your basket is empty, functionality disabled!");
             }
             return false;
         }
